Move token repositories to sub folder
