+++ src/User/TokenRepository/CacheRepository.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\User\Auth0;
+namespace CultuurNet\UDB3\User\TokenRepository;
 
 use CultuurNet\UDB3\User\ManagementToken;
 use DateTimeImmutable;
 use Doctrine\Common\Cache\Cache;
 
-class CacheRepository implements Auth0ManagementTokenRepository
+class CacheRepository implements TokenRepository
 {
     private const TOKEN_KEY = 'auth0_token';
